Added optional deletion (dequeueing) of select messages
The jms session is now in transacted mode to allow better control over
message browsing and the deletion of the message from the queue
afterwards. The deletion is achieved by consuming the messages selected
by the JMSMessageID.
We use session transaction, so in case of an error during processing,
all forwarded and consumed messages are rolled back.